requirements.txt
# local package
-e .

# external requirements
click
Sphinx
coverage
awscli
flake8
python-dotenv>=0.5.1

h5py==3.10.0
hydra-core==1.3.2
nibabel==5.2.0
nilearn==0.10.3
numpy==1.26.3
omegaconf==2.3.0
packaging==23.2
pandas==2.1.4
PyYAML==6.0.1
rich==13.7.0
scikit-learn==1.4.0
scipy==1.11.4
torch
torchvision
tqdm==4.66.1
wandb
